Issue #1404: DataStorage Acl tests are separated
