+++ api/src/test/java/com/epam/pipeline/acl/datastorage/DataStorageApiServiceTest.java
@@ -16,50 +16,20 @@
 
 package com.epam.pipeline.acl.datastorage;
 
-import com.epam.pipeline.controller.vo.data.storage.UpdateDataStorageItemVO;
 import com.epam.pipeline.controller.vo.security.EntityWithPermissionVO;
 import com.epam.pipeline.entity.datastorage.AbstractDataStorage;
-import com.epam.pipeline.entity.datastorage.AbstractDataStorageItem;
 import com.epam.pipeline.entity.datastorage.ContentDisposition;
-import com.epam.pipeline.entity.datastorage.DataStorageDownloadFileUrl;
-import com.epam.pipeline.entity.datastorage.DataStorageFile;
-import com.epam.pipeline.entity.datastorage.DataStorageItemContent;
 import com.epam.pipeline.entity.datastorage.DataStorageListing;
-import com.epam.pipeline.entity.datastorage.DataStorageStreamingContent;
-import com.epam.pipeline.entity.datastorage.DataStorageWithShareMount;
-import com.epam.pipeline.entity.datastorage.PathDescription;
-import com.epam.pipeline.entity.datastorage.StorageMountPath;
-import com.epam.pipeline.entity.datastorage.StorageUsage;
-import com.epam.pipeline.entity.datastorage.rules.DataStorageRule;
-import com.epam.pipeline.entity.pipeline.Pipeline;
 import com.epam.pipeline.entity.security.acl.AclClass;
-import com.epam.pipeline.manager.EntityManager;
-import com.epam.pipeline.manager.datastorage.DataStorageManager;
-import com.epam.pipeline.manager.datastorage.DataStorageRuleManager;
-import com.epam.pipeline.manager.datastorage.RunMountService;
-import com.epam.pipeline.manager.security.AuthManager;
-import com.epam.pipeline.manager.security.GrantPermissionManager;
 import com.epam.pipeline.security.UserContext;
 import com.epam.pipeline.security.acl.AclPermission;
-import com.epam.pipeline.test.acl.AbstractAclTest;
 import com.epam.pipeline.test.creator.datastorage.DatastorageCreatorUtils;
-import com.epam.pipeline.test.creator.pipeline.PipelineCreatorUtils;
 import com.epam.pipeline.test.creator.security.SecurityCreatorUtils;
 import org.junit.Test;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.authentication.TestingAuthenticationToken;
-import org.springframework.security.core.Authentication;
 import org.springframework.security.test.context.support.WithMockUser;
 
-import java.io.ByteArrayInputStream;
-import java.io.InputStream;
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-
 import static com.epam.pipeline.test.creator.CommonCreatorConstants.ID;
-import static com.epam.pipeline.test.creator.CommonCreatorConstants.ID_2;
 import static com.epam.pipeline.test.creator.CommonCreatorConstants.TEST_ARRAY;
 import static com.epam.pipeline.test.creator.CommonCreatorConstants.TEST_INT;
 import static com.epam.pipeline.test.creator.CommonCreatorConstants.TEST_STRING;
@@ -77,332 +47,7 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.times;
 import static org.mockito.Mockito.verify;
 
-public class DataStorageApiServiceTest extends AbstractAclTest {
-
-    private final AbstractDataStorage s3bucket = DatastorageCreatorUtils.getS3bucketDataStorage(ID, OWNER_USER);
-    private final AbstractDataStorage anotherS3bucket =
-            DatastorageCreatorUtils.getS3bucketDataStorage(ID_2, TEST_STRING);
-    private final DataStorageWithShareMount storageShareMount = DatastorageCreatorUtils.getDataStorageWithShareMount();
-    private final Authentication authentication = new TestingAuthenticationToken(new Object(), new Object());
-    private final UserContext context = SecurityCreatorUtils.getUserContext();
-    private final DataStorageListing dataStorageListing = DatastorageCreatorUtils.getDataStorageListing();
-    private final AbstractDataStorageItem dataStorageFile = DatastorageCreatorUtils.getDataStorageFile();
-    private final InputStream inputStream = new ByteArrayInputStream(TEST_STRING.getBytes());
-    private final DataStorageDownloadFileUrl downloadFileUrl = DatastorageCreatorUtils.getDataStorageDownloadFileUrl();
-    private final Pipeline pipeline = PipelineCreatorUtils.getPipeline(OWNER_USER);
-    private final DataStorageRule dataStorageRule = DatastorageCreatorUtils.getDataStorageRule();
-    private final DataStorageItemContent dataStorageItemContent =
-            DatastorageCreatorUtils.getDefaultDataStorageItemContent();
-    private final DataStorageStreamingContent dataStorageStreamingContent =
-            DatastorageCreatorUtils.getDefaultDataStorageStreamingContent(inputStream);
-    private final List<PathDescription> pathDescriptionList = DatastorageCreatorUtils.getPathDescriptionList();
-    private final StorageUsage storageUsage = DatastorageCreatorUtils.getStorageUsage();
-    private final StorageMountPath storageMountPath = DatastorageCreatorUtils.getStorageMountPath();
-
-    private final List<String> testList = Collections.singletonList(TEST_STRING);
-    private final List<UpdateDataStorageItemVO> dataStorageItemVOList =
-            DatastorageCreatorUtils.getUpdateDataStorageItemVOList();
-    private final List<DataStorageFile> dataStorageFileList = DatastorageCreatorUtils.getDataStorageFileList();
-    private final List<DataStorageDownloadFileUrl> downloadFileUrlList =
-            DatastorageCreatorUtils.getDataStorageDownloadFileUrlList();
-    private final List<DataStorageRule> dataStorageRuleList = DatastorageCreatorUtils.getDataStorageRuleList();
-
-    @Autowired
-    private DataStorageApiService dataStorageApiService;
-
-    @Autowired
-    private GrantPermissionManager grantPermissionManager;
-
-    @Autowired
-    private DataStorageManager mockDataStorageManager;
-
-    @Autowired
-    private DataStorageRuleManager mockDataStorageRuleManager;
-
-    @Autowired
-    private AuthManager mockAuthManager;
-
-    @Autowired
-    private EntityManager mockEntityManager;
-
-    @Autowired
-    private RunMountService mockRunMountService;
-
-    @Test
-    @WithMockUser(roles = ADMIN_ROLE)
-    public void shouldReturnDataStoragesForAdmin() {
-        doReturn(mutableListOf(s3bucket)).when(mockDataStorageManager).getDataStorages();
-
-        assertThat(dataStorageApiService.getDataStorages()).hasSize(1).contains(s3bucket);
-    }
-
-    @Test
-    @WithMockUser(username = SIMPLE_USER)
-    public void shouldReturnDataStoragesWhenPermissionIsGranted() {
-        initAclEntity(s3bucket, AclPermission.READ);
-        doReturn(mutableListOf(s3bucket)).when(mockDataStorageManager).getDataStorages();
-
-        List<AbstractDataStorage> returnedDataStorages = dataStorageApiService.getDataStorages();
-
-        assertThat(dataStorageApiService.getDataStorages()).hasSize(1).contains(s3bucket);
-        assertThat(returnedDataStorages.get(0).getMask()).isEqualTo(1);
-    }
-
-    @Test
-    @WithMockUser(username = SIMPLE_USER)
-    public void shouldReturnDataStoragesWhichPermissionIsGranted() {
-        initAclEntity(s3bucket, AclPermission.READ);
-        initAclEntity(anotherS3bucket, AclPermission.NO_READ);
-        doReturn(mutableListOf(s3bucket, anotherS3bucket)).when(mockDataStorageManager).getDataStorages();
-
-        assertThat(dataStorageApiService.getDataStorages()).hasSize(1).contains(s3bucket);
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldReturnEmptyDataStorageListWhenPermissionIsNotGranted() {
-        initAclEntity(s3bucket, AclPermission.NO_READ);
-        doReturn(mutableListOf(s3bucket)).when(mockDataStorageManager).getDataStorages();
-
-        assertThat(dataStorageApiService.getDataStorages()).isEmpty();
-    }
-
-    @Test
-    @WithMockUser(roles = ADMIN_ROLE)
-    public void shouldReturnWritableDataStoragesForAdmin() {
-        doReturn(mutableListOf(s3bucket)).when(mockDataStorageManager).getDataStorages();
-
-        assertThat(dataStorageApiService.getWritableStorages()).hasSize(1).contains(s3bucket);
-    }
-
-    @Test
-    @WithMockUser(username = SIMPLE_USER)
-    public void shouldReturnWritableDataStoragesWhenPermissionIsGranted() {
-        initAclEntity(s3bucket, Arrays.asList(new UserPermission(SIMPLE_USER, AclPermission.READ.getMask()),
-                new UserPermission(SIMPLE_USER, AclPermission.WRITE.getMask())));
-        doReturn(mutableListOf(s3bucket)).when(mockDataStorageManager).getDataStorages();
-
-        List<AbstractDataStorage> returnedDataStorages = dataStorageApiService.getWritableStorages();
-
-        assertThat(returnedDataStorages).hasSize(1).contains(s3bucket);
-        assertThat(returnedDataStorages.get(0).getMask()).isEqualTo(3);
-    }
-
-    @Test
-    @WithMockUser(username = SIMPLE_USER)
-    public void shouldReturnWritableDataStoragesWhichPermissionIsGranted() {
-        initAclEntity(s3bucket, Arrays.asList(new UserPermission(SIMPLE_USER, AclPermission.READ.getMask()),
-                new UserPermission(SIMPLE_USER, AclPermission.WRITE.getMask())));
-        initAclEntity(anotherS3bucket, Arrays.asList(new UserPermission(SIMPLE_USER, AclPermission.NO_READ.getMask()),
-                new UserPermission(SIMPLE_USER, AclPermission.WRITE.getMask())));
-        doReturn(mutableListOf(s3bucket, anotherS3bucket)).when(mockDataStorageManager).getDataStorages();
-
-        assertThat(dataStorageApiService.getWritableStorages()).hasSize(1).contains(s3bucket);
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldReturnEmptyWritableDataStorageListWhenPermissionIsNotGranted() {
-        initAclEntity(s3bucket, Arrays.asList(new UserPermission(SIMPLE_USER, AclPermission.NO_READ.getMask()),
-                new UserPermission(SIMPLE_USER, AclPermission.NO_WRITE.getMask())));
-        doReturn(mutableListOf(s3bucket)).when(mockDataStorageManager).getDataStorages();
-
-        assertThat(dataStorageApiService.getWritableStorages()).isEmpty();
-    }
-
-    @Test
-    @WithMockUser(roles = ADMIN_ROLE)
-    public void shouldReturnAvailableStoragesWithShareMountForAdmin() {
-        doReturn(mutableListOf(storageShareMount)).when(mockDataStorageManager).getDataStoragesWithShareMountObject(ID);
-
-        assertThat(dataStorageApiService.getAvailableStoragesWithShareMount(ID)).hasSize(1).contains(storageShareMount);
-    }
-
-    @Test
-    @WithMockUser(username = SIMPLE_USER)
-    public void shouldReturnAvailableStoragesWithShareMountWhenReadPermissionIsGranted() {
-        initAclEntity(s3bucket, AclPermission.READ);
-        doReturn(mutableListOf(storageShareMount)).when(mockDataStorageManager).getDataStoragesWithShareMountObject(ID);
-
-        assertThat(dataStorageApiService.getAvailableStoragesWithShareMount(ID)).hasSize(1).contains(storageShareMount);
-    }
-
-    @Test
-    @WithMockUser(username = SIMPLE_USER)
-    public void shouldReturnAvailableStoragesWithShareMountWhenWritePermissionIsGranted() {
-        initAclEntity(s3bucket, AclPermission.WRITE);
-        doReturn(mutableListOf(storageShareMount)).when(mockDataStorageManager).getDataStoragesWithShareMountObject(ID);
-
-        assertThat(dataStorageApiService.getAvailableStoragesWithShareMount(ID)).hasSize(1).contains(storageShareMount);
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldReturnEmptyAvailableStoragesWithShareMountWhenPermissionIsNotGranted() {
-        initAclEntity(s3bucket);
-        doReturn(mutableListOf(storageShareMount)).when(mockDataStorageManager).getDataStoragesWithShareMountObject(ID);
-
-        assertThat(dataStorageApiService.getAvailableStoragesWithShareMount(ID)).isEmpty();
-    }
-
-    @Test
-    @WithMockUser(roles = ADMIN_ROLE)
-    public void shouldReturnAvailableStoragesForAdmin() {
-        doReturn(mutableListOf(s3bucket)).when(mockDataStorageManager).getDataStorages();
-
-        assertThat(dataStorageApiService.getAvailableStorages()).hasSize(1).contains(s3bucket);
-    }
-
-    @Test
-    @WithMockUser(username = SIMPLE_USER)
-    public void shouldReturnAvailableStoragesWhenReadPermissionIsGranted() {
-        initAclEntity(s3bucket, AclPermission.READ);
-        doReturn(mutableListOf(s3bucket)).when(mockDataStorageManager).getDataStorages();
-
-        List<AbstractDataStorage> returnedDataStorages = dataStorageApiService.getAvailableStorages();
-
-        assertThat(returnedDataStorages).hasSize(1).contains(s3bucket);
-        assertThat(returnedDataStorages.get(0).getMask()).isEqualTo(1);
-    }
-
-    @Test
-    @WithMockUser(username = SIMPLE_USER)
-    public void shouldReturnAvailableStoragesWhenWritePermissionIsGranted() {
-        initAclEntity(s3bucket, AclPermission.WRITE);
-        doReturn(mutableListOf(s3bucket)).when(mockDataStorageManager).getDataStorages();
-
-        List<AbstractDataStorage> returnedDataStorages = dataStorageApiService.getAvailableStorages();
-
-        assertThat(returnedDataStorages).hasSize(1).contains(s3bucket);
-        assertThat(returnedDataStorages.get(0).getMask()).isEqualTo(2);
-    }
-
-    @Test
-    @WithMockUser(username = SIMPLE_USER)
-    public void shouldReturnAvailableStoragesWhichPermissionIsGranted() {
-        initAclEntity(s3bucket, AclPermission.READ);
-        initAclEntity(anotherS3bucket);
-        doReturn(mutableListOf(s3bucket, anotherS3bucket)).when(mockDataStorageManager).getDataStorages();
-
-        List<AbstractDataStorage> returnedDataStorages = dataStorageApiService.getAvailableStorages();
-
-        assertThat(returnedDataStorages).hasSize(1).contains(s3bucket);
-        assertThat(returnedDataStorages.get(0).getMask()).isEqualTo(1);
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldReturnEmptyAvailableStoragesWhenWritePermissionIsNotGranted() {
-        initAclEntity(s3bucket);
-        doReturn(mutableListOf(s3bucket)).when(mockDataStorageManager).getDataStorages();
-
-        assertThat(dataStorageApiService.getAvailableStorages()).isEmpty();
-    }
-
-    @Test
-    @WithMockUser(roles = ADMIN_ROLE)
-    public void shouldLoadDataStorageForAdmin() {
-        final UserContext context = SecurityCreatorUtils.getUserContext();
-        doReturn(s3bucket).when(mockDataStorageManager).load(ID);
-        initMocks(OWNER_USER, context);
-
-        assertThat(dataStorageApiService.load(ID)).isEqualTo(s3bucket);
-    }
-
-    @Test
-    @WithMockUser(username = SIMPLE_USER)
-    public void shouldLoadDataStorageWhenPermissionIsGranted() {
-        final UserContext context = SecurityCreatorUtils.getUserContext();
-        initAclEntity(s3bucket, AclPermission.READ);
-        doReturn(s3bucket).when(mockDataStorageManager).load(ID);
-        initMocks(OWNER_USER, context);
-
-        AbstractDataStorage returnedDataStorage = dataStorageApiService.load(ID);
-
-        assertThat(returnedDataStorage).isEqualTo(s3bucket);
-        assertThat(returnedDataStorage.getMask()).isEqualTo(1);
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldDenyLoadDataStorageWhenStoragePermissionIsNotGranted() {
-        initAclEntity(s3bucket, AclPermission.NO_READ);
-        doReturn(s3bucket).when(mockDataStorageManager).load(ID);
-        mockS3bucket();
-        mockAuthUser(SIMPLE_USER);
-
-        assertThrows(AccessDeniedException.class, () -> dataStorageApiService.load(ID));
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldDenyLoadDataStorageWhenCheckStorageSharedPermissionIsNotGranted() {
-        final UserContext context = SecurityCreatorUtils.getUserContext(true);
-        final AbstractDataStorage s3bucket = DatastorageCreatorUtils.getS3bucketDataStorage(ID, OWNER_USER, false);
-        initAclEntity(s3bucket, AclPermission.READ);
-        doReturn(s3bucket).when(mockDataStorageManager).load(ID);
-        initMocks(OWNER_USER, context);
-
-        assertThrows(AccessDeniedException.class, () -> dataStorageApiService.load(ID));
-    }
-
-    @Test
-    @WithMockUser(roles = ADMIN_ROLE)
-    public void shouldLoadByNameOrIdForAdmin() {
-        doReturn(s3bucket).when(mockDataStorageManager).loadByNameOrId(TEST_STRING);
-
-        assertThat(dataStorageApiService.loadByNameOrId(TEST_STRING)).isEqualTo(s3bucket);
-    }
-
-    @Test
-    @WithMockUser(username = SIMPLE_USER)
-    public void shouldLoadByNameOrIdWhenPermissionIsGranted() {
-        initAclEntity(s3bucket, AclPermission.READ);
-        doReturn(s3bucket).when(mockDataStorageManager).loadByNameOrId(TEST_STRING);
-
-        AbstractDataStorage returnedDataStorage = dataStorageApiService.loadByNameOrId(TEST_STRING);
-
-        assertThat(returnedDataStorage).isEqualTo(s3bucket);
-        assertThat(returnedDataStorage.getMask()).isEqualTo(1);
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldDenyLoadingByNameOrIdWhenPermissionIsNotGranted() {
-        initAclEntity(s3bucket, AclPermission.NO_READ);
-        doReturn(s3bucket).when(mockDataStorageManager).loadByNameOrId(TEST_STRING);
-
-        assertThrows(AccessDeniedException.class, () -> dataStorageApiService.loadByNameOrId(TEST_STRING));
-    }
-
-    @Test
-    @WithMockUser(roles = ADMIN_ROLE)
-    public void shouldLoadByPathOrIdForAdmin() {
-        doReturn(s3bucket).when(mockDataStorageManager).loadByPathOrId(TEST_STRING);
-
-        assertThat(dataStorageApiService.loadByPathOrId(TEST_STRING)).isEqualTo(s3bucket);
-    }
-
-    @Test
-    @WithMockUser(username = SIMPLE_USER)
-    public void shouldLoadByPathOrIdWhenPermissionIsGranted() {
-        initAclEntity(s3bucket, AclPermission.READ);
-        doReturn(s3bucket).when(mockDataStorageManager).loadByPathOrId(TEST_STRING);
-
-        AbstractDataStorage returnedDataStorage = dataStorageApiService.loadByPathOrId(TEST_STRING);
-
-        assertThat(returnedDataStorage).isEqualTo(s3bucket);
-        assertThat(returnedDataStorage.getMask()).isEqualTo(1);
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldDenyLoadingByPathOrIdWhenPermissionIsNotGranted() {
-        initAclEntity(s3bucket, AclPermission.NO_READ);
-        doReturn(s3bucket).when(mockDataStorageManager).loadByPathOrId(TEST_STRING);
-
-        assertThrows(AccessDeniedException.class, () -> dataStorageApiService.loadByPathOrId(TEST_STRING));
-    }
+public class DataStorageApiServiceTest extends AbstractDataStorageAclTest {
 
     @Test
     @WithMockUser(roles = ADMIN_ROLE)
@@ -1068,58 +713,6 @@ public class DataStorageApiServiceTest extends AbstractAclTest {
 
     @Test
     @WithMockUser(roles = ADMIN_ROLE)
-    public void shouldLoadRulesForAdmin() {
-        doReturn(dataStorageRuleList).when(mockDataStorageRuleManager).loadRules(ID, TEST_STRING);
-
-        assertThat(dataStorageApiService.loadRules(ID, TEST_STRING)).isEqualTo(dataStorageRuleList);
-    }
-
-    @Test
-    @WithMockUser(username = SIMPLE_USER)
-    public void shouldLoadRulesWhenPermissionIsGranted() {
-        initAclEntity(pipeline, AclPermission.READ);
-        doReturn(dataStorageRuleList).when(mockDataStorageRuleManager).loadRules(ID, TEST_STRING);
-
-        assertThat(dataStorageApiService.loadRules(ID, TEST_STRING)).isEqualTo(dataStorageRuleList);
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldDenyLoadRulesWhenPermissionIsNotGranted() {
-        initAclEntity(pipeline);
-        doReturn(dataStorageRuleList).when(mockDataStorageRuleManager).loadRules(ID, TEST_STRING);
-
-        assertThrows(AccessDeniedException.class, () -> dataStorageApiService.loadRules(ID, TEST_STRING));
-    }
-
-    @Test
-    @WithMockUser(roles = ADMIN_ROLE)
-    public void shouldDeleteRuleForAdmin() {
-        doReturn(dataStorageRule).when(mockDataStorageRuleManager).deleteRule(ID, TEST_STRING);
-
-        assertThat(dataStorageApiService.deleteRule(ID, TEST_STRING)).isEqualTo(dataStorageRule);
-    }
-
-    @Test
-    @WithMockUser(username = SIMPLE_USER)
-    public void shouldDeleteRuleWhenPermissionIsGranted() {
-        initAclEntity(pipeline, AclPermission.WRITE);
-        doReturn(dataStorageRule).when(mockDataStorageRuleManager).deleteRule(ID, TEST_STRING);
-
-        assertThat(dataStorageApiService.deleteRule(ID, TEST_STRING)).isEqualTo(dataStorageRule);
-    }
-
-    @Test
-    @WithMockUser
-    public void shouldDenyDeleteRuleWhenPermissionIsNotGranted() {
-        initAclEntity(pipeline);
-        doReturn(dataStorageRule).when(mockDataStorageRuleManager).deleteRule(ID, TEST_STRING);
-
-        assertThrows(AccessDeniedException.class, () -> dataStorageApiService.deleteRule(ID, TEST_STRING));
-    }
-
-    @Test
-    @WithMockUser(roles = ADMIN_ROLE)
     public void shouldUpdateObjectTagsForAdmin() {
         doReturn(TEST_STRING_MAP).when(mockDataStorageManager)
                 .updateDataStorageObjectTags(ID, TEST_STRING, TEST_STRING_MAP, TEST_STRING, true);
@@ -1694,20 +1287,5 @@ public class DataStorageApiServiceTest extends AbstractAclTest {
 
         assertThrows(AccessDeniedException.class, () ->
                 dataStorageApiService.getStorageUsage(TEST_STRING, TEST_STRING));
-    }
-
-    private void initMocks(String user, UserContext context) {
-        mockAuthUser(user);
-        mockS3bucket();
-        doReturn(context).when(mockAuthManager).getUserContext();
-    }
-
-    private void mockAuthUser(String user) {
-        doReturn(user).when(mockAuthManager).getAuthorizedUser();
-        doReturn(authentication).when(mockAuthManager).getAuthentication();
-    }
-
-    private void mockS3bucket() {
-        doReturn(s3bucket).when(mockEntityManager).load(AclClass.DATA_STORAGE, ID);
     }
 }
