+++ api/src/test/java/com/epam/pipeline/test/acl/AbstractAclTest.java
@@ -54,6 +54,7 @@ public abstract class AbstractAclTest {
     protected static final String OWNER_USER = "OWNER";
     protected static final String GENERAL_USER_ROLE = "USER";
     protected static final String SIMPLE_USER_ROLE = "SIMPLE_USER";
+    protected static final String STORAGE_MANAGER_ROLE = "STORAGE_MANAGER";
     protected static final String SIMPLE_USER = "SIMPLE_USER";
     protected static final String ANOTHER_SIMPLE_USER = "ANOTHER_SIMPLE_USER";
     protected static final String TEST_NAME = "TEST_NAME";
